[department-of-veterans-affairs/va-virtual-agent#59] Extracts Feature Toggle loading into parent component of webchat
